Extract awaitPromises helper and add asyncMap; refactor heating insights
- Move awaitPromises out of device-helpers.ts into helpers/promises.ts so
  it can be shared without a routes dependency.
- Add asyncMap to helpers/array.ts (Promise.all over a mapped array).
- Rewrite the /insights/heating handler to use both helpers, replacing
  nested Promise.all destructures with named awaitPromises objects.
- Make HeatingInsightsApiResponse.heatPump non-nullable; there is always
  a heat pump, so the |null guard was misleading.
