+++ server/src/routes/api/device-helpers.ts
@@ -1,18 +1,9 @@
 import { Device, NumericEvent, BooleanEvent } from '../../models';
 import { NumericEventApiResponse, BooleanEventApiResponse, EnumEventApiResponse, RestDeviceResponse, CapabilityApiResponse } from '../../api/types';
 import dayjs from '../../dayjs';
+import { awaitPromises } from '../../helpers/promises';
 
-type AwaitedObject<T> = {
-  [K in keyof T]: T[K] extends Promise<infer U> ? U : T[K];
-};
-
-export async function awaitPromises<T extends Record<string, unknown>>(obj: T): Promise<AwaitedObject<T>> {
-  const entries = await Promise.all(
-    Object.entries(obj).map(async ([key, value]) => [key, await value])
-  );
-
-  return Object.fromEntries(entries) as AwaitedObject<T>;
-}
+export { awaitPromises };
 
 export function mapNumericEvent(eventPromise: Promise<NumericEvent | null>): Promise<NumericEventApiResponse> {
   return eventPromise.then(event => {
